requirements.txt
asgiref==3.5.2
Django==4.1.3
sqlparse==0.4.3
tzdata==2022.6

django-mptt~=0.14.0